[red] parse number array
